fix broken unit tests
